data-generator/requirements.txt
kafka-python==2.0.2
faker==19.13.0
python-dotenv==1.0.0 